refactor: move device
