+++ providers/auth-provider/login.ts
@@ -1,9 +1,10 @@
 import { Platform } from 'react-native';
 
-import { getDeviceName, getLoginId } from '../utils';
+import { getLoginId } from '../utils';
 
 import { getTokenKey } from '~/config';
 import { TwoFactorChallengeError } from '~/errors/TwoFactorChallengeError';
+import { getDeviceName } from '~/lib/device';
 import { getHttp } from '~/lib/http';
 import { setItemAsync } from '~/lib/store';
