+++ providers/auth-provider/two-factor-challenge.ts
@@ -1,9 +1,10 @@
 import { Platform } from 'react-native';
 
-import { getDeviceName, getLoginId } from '../utils';
+import { getLoginId } from '../utils';
 
 import { getTokenKey } from '~/config';
 import { EmailUnavailableError } from '~/errors/EmailUnavailableError';
+import { getDeviceName } from '~/lib/device';
 import { getHttp } from '~/lib/http';
 import { deleteItemAsync, getItemAsync, setItemAsync } from '~/lib/store';
